requirements.txt
pandas~=2.2.2
ccxt~=4.3.50
discord.py
ta~=0.11.0
pytest~=8.2.2
pytest-mock~=3.14.0
pytest-asyncio~=0.23.7
numpy~=2.0.0